requirements.txt
fastapi
uvicorn
requests
sounddevice
numpy
colorama
google-cloud-speech
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
